// src/components/custom-toast/notification-types.ts
export type NotificationType =
  | 'PAYOUT'
  | 'COLLECTION'
  | 'COLLECTION_RECEIVED'
  | 'COLLECTION_PROCESSED'
  | 'COLLECTION_USDT_PROCESSED'
  | 'PAYOUT_PROCESSED'
  | 'PAYOUT_USDT_PROCESSED'
  | 'POOL_BALANCE'
  | 'ACCESS_TO_MERCHANT_PORTAL'
  | 'PAYOUT_IDR_PROCESSED'
  | 'COLLECTION_IDR_PROCESSED'
  | 'PAYOUT_THB_PROCESSED'
  | 'COLLECTION_THB_PROCESSED'
  | 'COLLECTION_BEP20_PROCESSED'
  | 'PAYOUT_BEP20_PROCESSED'

export const notificationTypeColors: Record<NotificationType, string> = {
  PAYOUT: '#334155',
  COLLECTION: '#14532D',
  COLLECTION_RECEIVED: '#0F766E',
  COLLECTION_PROCESSED: '#1D4ED8',
  COLLECTION_USDT_PROCESSED: '#047857',
  PAYOUT_PROCESSED: '#4338CA',
  PAYOUT_USDT_PROCESSED: '#065F46',
  POOL_BALANCE: '#5B21B6',
  ACCESS_TO_MERCHANT_PORTAL: '#1E3A5F',
  PAYOUT_IDR_PROCESSED: '#6D28D9',
  COLLECTION_IDR_PROCESSED: '#0F766E',
  PAYOUT_THB_PROCESSED: '#9A3412',
  COLLECTION_THB_PROCESSED: '#B45309',
  COLLECTION_BEP20_PROCESSED: '#0F766E',
  PAYOUT_BEP20_PROCESSED: '#155E75'
}